# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vinted clone API specification
 * OpenAPI spec version: 0.1.0
 */
export type AuthUserSellerStatus = typeof AuthUserSellerStatus[keyof typeof AuthUserSellerStatus];


export const AuthUserSellerStatus = {
  none: 'none',
  pending: 'pending',
  approved: 'approved',
} as const;

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  sellerStatus: AuthUserSellerStatus;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ItemCondition = typeof ItemCondition[keyof typeof ItemCondition];


export const ItemCondition = {
  New_with_tags: 'New with tags',
  Like_new: 'Like new',
  Good: 'Good',
  Fair: 'Fair',
} as const;

export interface Item {
  id: string;
  title: string;
  brand: string;
  price: number;
  originalPrice?: number | null;
  size: string;
  condition: ItemCondition;
  category: string;
  description: string;
  color?: string | null;
  sellerId: string;
  likesCount: number;
  viewsCount: number;
  images: string[];
  createdAt: string;
}

export interface User {
  id: string;
  name: string;
  bio?: string | null;
  rating: number;
  reviewCount: number;
  itemCount: number;
  followersCount: number;
  followingCount: number;
  verified: boolean;
  createdAt: string;
}

export type ItemDetail = Item & {
  seller: User;
};

export interface ItemListResponse {
  items: Item[];
  total: number;
  page: number;
  limit: number;
  hasMore: boolean;
}

export type CreateItemBodyCondition = typeof CreateItemBodyCondition[keyof typeof CreateItemBodyCondition];


export const CreateItemBodyCondition = {
  New_with_tags: 'New with tags',
  Like_new: 'Like new',
  Good: 'Good',
  Fair: 'Fair',
} as const;

export interface CreateItemBody {
  /** @minLength 2 */
  title: string;
  /** @minLength 1 */
  brand: string;
  /** @minimum 0.01 */
  price: number;
  originalPrice?: number | null;
  size: string;
  condition: CreateItemBodyCondition;
  category: string;
  description: string;
  color?: string | null;
  sellerId: string;
  images: string[];
}

export interface ViewCountResponse {
  viewsCount: number;
}

export interface LikeBody {
  userId: string;
}

export interface LikeResponse {
  liked: boolean;
  likesCount: number;
}

export interface CreateUserBody {
  /** @minLength 2 */
  name: string;
  bio?: string | null;
}

export interface Category {
  id: string;
  label: string;
  icon: string;
}

export interface Conversation {
  id: string;
  buyerId: string;
  sellerId: string;
  itemId?: string | null;
  lastMessage?: string | null;
  lastMessageAt?: string | null;
  unreadCount: number;
  otherUser?: User;
  item?: Item | null;
  createdAt: string;
}

export interface CreateConversationBody {
  buyerId: string;
  sellerId: string;
  itemId?: string | null;
  initialMessage?: string;
}

export interface Message {
  id: string;
  conversationId: string;
  senderId: string;
  text: string;
  createdAt: string;
}

export interface SendMessageBody {
  senderId: string;
  /** @minLength 1 */
  text: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListItemsParams = {
category?: string;
q?: string;
size?: string;
condition?: string;
minPrice?: number;
maxPrice?: number;
page?: number;
limit?: number;
};

export type GetUserItemsParams = {
page?: number;
limit?: number;
};

export type ListConversationsParams = {
userId: string;
};