# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vinted clone API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: items
    description: Marketplace listings
  - name: users
    description: User profiles
  - name: categories
    description: Item categories
  - name: conversations
    description: Messaging between buyers and sellers

paths:
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success or /api/login on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List items
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - name: size
          in: query
          schema: { type: string }
        - name: condition
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"
    post:
      operationId: createItem
      tags: [items]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemBody"
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Item found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /items/{id}/view:
    post:
      operationId: trackItemView
      tags: [items]
      summary: Increment view count
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: View tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewCountResponse"

  /items/{id}/like:
    post:
      operationId: likeItem
      tags: [items]
      summary: Like an item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LikeBody"
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}/items:
    get:
      operationId: getUserItems
      tags: [users]
      summary: Get items listed by a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: User items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemListResponse"

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List conversations for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Start a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [conversations]
      summary: Get messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [conversations]
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string

  schemas:
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, sellerStatus]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        sellerStatus:
          type: string
          enum: [none, pending, approved]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        brand:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        size:
          type: string
        condition:
          type: string
          enum: [New with tags, Like new, Good, Fair]
        category:
          type: string
        description:
          type: string
        color:
          type: string
          nullable: true
        sellerId:
          type: string
          format: uuid
        likesCount:
          type: integer
        viewsCount:
          type: integer
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, brand, price, size, condition, category, description, sellerId, likesCount, viewsCount, images, createdAt]

    ItemDetail:
      allOf:
        - $ref: "#/components/schemas/Item"
        - type: object
          properties:
            seller:
              $ref: "#/components/schemas/User"
          required: [seller]

    ItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean
      required: [items, total, page, limit, hasMore]

    CreateItemBody:
      type: object
      properties:
        title:
          type: string
          minLength: 2
        brand:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0.01
        originalPrice:
          type: number
          nullable: true
        size:
          type: string
        condition:
          type: string
          enum: [New with tags, Like new, Good, Fair]
        category:
          type: string
        description:
          type: string
        color:
          type: string
          nullable: true
        sellerId:
          type: string
          format: uuid
        images:
          type: array
          items:
            type: string
      required: [title, brand, price, size, condition, category, description, sellerId, images]

    ViewCountResponse:
      type: object
      properties:
        viewsCount:
          type: integer
      required: [viewsCount]

    LikeBody:
      type: object
      properties:
        userId:
          type: string
          format: uuid
      required: [userId]

    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
      required: [liked, likesCount]

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        bio:
          type: string
          nullable: true
        rating:
          type: number
        reviewCount:
          type: integer
        itemCount:
          type: integer
        followersCount:
          type: integer
        followingCount:
          type: integer
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, rating, reviewCount, itemCount, followersCount, followingCount, verified, createdAt]

    CreateUserBody:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        bio:
          type: string
          nullable: true
      required: [name]

    Category:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        icon:
          type: string
      required: [id, label, icon]

    Conversation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        buyerId:
          type: string
          format: uuid
        sellerId:
          type: string
          format: uuid
        itemId:
          type: string
          format: uuid
          nullable: true
        lastMessage:
          type: string
          nullable: true
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        unreadCount:
          type: integer
        otherUser:
          $ref: "#/components/schemas/User"
        item:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Item"
        createdAt:
          type: string
          format: date-time
      required: [id, buyerId, sellerId, unreadCount, createdAt]

    CreateConversationBody:
      type: object
      properties:
        buyerId:
          type: string
          format: uuid
        sellerId:
          type: string
          format: uuid
        itemId:
          type: string
          format: uuid
          nullable: true
        initialMessage:
          type: string
      required: [buyerId, sellerId]

    Message:
      type: object
      properties:
        id:
          type: string
          format: uuid
        conversationId:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        text:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, conversationId, senderId, text, createdAt]

    SendMessageBody:
      type: object
      properties:
        senderId:
          type: string
          format: uuid
        text:
          type: string
          minLength: 1
      required: [senderId, text]